Pipe Current State into Target State
This change pipes CurrentHeatingCoolingState into TargetHeatingCoolingState; as opposed to the other way around in the previous commit. Error was not noticeable until I started work on functionality for having two constant zones as two thermostats (not in this file). 

Also added '* )' case to fan accessory for characteristic 'On' for cases where '$mode' returns something unexpected; just like is currently done for 'CurrentHeatingCoolingState/TargetHeatingCoolingState' characteristic for the thermostat accessory.