Rewrite of code.
Comments added and a new approach to how to handle two constant zones as individual thermostats that interact with each other. Now there is working conditional logic to check if they other constant zone is open or closed before it decides to shut the zone or shut off the master control unit itself.