cmd4 v3.0.x required config.json
Requires "restartRecover" set to 'false' to delete previous accessory cache on first restart; can be set to true after this to reset the cache every restart of homebridge. Added new polling method thermostat and temperature sensor accessories and "outputConstants" added to each accessory and set as required (true or false).